+++ ElvUI_LuckyoneUI/Profiles/BigWigs.lua
@@ -9,8 +9,8 @@ function L1UI:Setup_BigWigs(layout)
 	if not E:IsAddOnEnabled('BigWigs') then return end
 
 	-- Profile names
-	local name = L1UI.Me and 'Luckyone Main' or 'Luckyone Main '..L1UI.Version
-	local name_healing = L1UI.Me and 'Luckyone Healing' or 'Luckyone Healing '..L1UI.Version
+	local name = E.global.L1UI.dev and 'Luckyone Main' or 'Luckyone Main '..L1UI.Version
+	local name_healing = E.global.L1UI.dev and 'Luckyone Healing' or 'Luckyone Healing '..L1UI.Version
 
 	-- Required to add profiles to BigWigs3DB
 	if not IsAddOnLoaded('BigWigs_Core') then LoadAddOn('BigWigs_Core') end