+++ ElvUI_LuckyoneUI/Profiles/Quartz.lua
@@ -12,8 +12,8 @@ function L1UI:Setup_Quartz(layout)
 	E.db.unitframe.units.target.castbar.enable = false
 
 	-- Profile names
-	local name = L1UI.Me and 'Luckyone Main' or 'Luckyone Main '..L1UI.Version
-	local name_healing = L1UI.Me and 'Luckyone Healing' or 'Luckyone Healing '..L1UI.Version
+	local name = E.global.L1UI.dev and 'Luckyone Main' or 'Luckyone Main '..L1UI.Version
+	local name_healing = E.global.L1UI.dev and 'Luckyone Healing' or 'Luckyone Healing '..L1UI.Version
 
 	-- Disable LibDualSpec to set the profile
 	Quartz3DB["namespaces"]["LibDualSpec-1.0"] = Quartz3DB["namespaces"]["LibDualSpec-1.0"] or {}