+++ ElvUI_LuckyoneUI/Settings/Profile.lua
@@ -2,5 +2,8 @@ local L1UI, E, L, V, P, G = unpack(select(2, ...))
 
 -- Defaults: E.db.L1UI
 P.L1UI = {
+	general = {
+		auto_update = false,
+	},
 	install_version = nil,
 }