+++ ElvUI_LuckyoneUI/Init.lua
@@ -4,6 +4,7 @@ local EP = LibStub('LibElvUIPlugin-1.0')
 local PI = E:GetModule('PluginInstaller')
 
 local _G = _G
+local tostring = tostring
 local GetAddOnMetadata = GetAddOnMetadata
 
 local addon, Engine = ...